20220224/push_subscriber/requirements.txt
flask
google-cloud-firestore
google-cloud-pubsub
gunicorn
numpy
